Complete Q-058 callable recovery checkpoint

+++ decomp/ST.exe/functions/004D9820/decomp.c
@@ -44,6 +44,7 @@ undefined4 __fastcall FUN_004d9820(int *param_1)
                         (this != (STWorldObject *)param_1)))))) &&
                      /* ST_CALLSITE[004D9925]: CALL dword ptr [EDX + 0x2c] */
                      (iVar3 = this->GetObjectTypeId(), iVar3 == 0x3a)))) &&
+
                    (iVar2 = FUN_006aced8(iVar9,iVar7,param_1[0x16c],param_1[0x16d]), iVar2 < 0xe)) {
                   param_1[0x137] = param_1[0x137] + 1;
                   param_1[0x138] = param_1[0x138] + iVar2;
@@ -65,26 +66,33 @@ undefined4 __fastcall FUN_004d9820(int *param_1)
     iVar7 = 0;
   }
   if (param_1[0x137] == 0) {
+
     iVar6 = thunk_FUN_004d8230();
   }
   else {
     lVar10 = FUN_006b12cc(0xd);
     iVar9 = (int)lVar10;
     lVar10 = FUN_006b12cc(iVar7);
+
     iVar7_mg3 = FUN_006b12a8((int)lVar10,iVar9);
     iVar7 = iVar7_mg3 + DAT_007eda8c;
+
     iVar9 = thunk_FUN_004d8230();
     lVar10 = FUN_006b12cc(iVar9);
+
     iVar7_mg4 = FUN_006b12dc((int)lVar10,iVar7);
+
     iVar6 = FUN_006b1280(iVar7_mg4);
   }
   param_1[0x139] = iVar6;
   if ((uint)(param_1[0x139] + param_1[0x135]) <= g_playSystem_00802A38->field_00E4) {
     param_1[0x135] = g_playSystem_00802A38->field_00E4;
     param_1[0x134] = param_1[0x134] + 2;
+
     thunk_FUN_004d82b0(2);
     if (99 < param_1[0x134]) {
-      thunk_FUN_004d9b20((int)param_1);
+
+      thunk_FUN_004d9b20(param_1);
     }
   }
   return 0;
