+++ decomp/ST.exe/functions/0067B210/decomp.c
@@ -5,7 +5,7 @@ void __thiscall FUN_0067b210(void *this,int param_1,int param_2)
 
 {
   if ((-1 < param_1) && (param_1 < 3)) {
-    thunk_FUN_0067b260((int)this);
+    thunk_FUN_0067b260(this);
     STField<undefined1>(this,0x662) = 1;
     STField<short>(this,0x660) = (short)param_1;
     if (param_2 < 1) {
